Modificando a classe Batalha.java
